add model/message
