+++ rocketmq-go/model/pull_result.go
