--- rocketmq-go/pull_status.go
@@ -1,10 +1,0 @@
-package rocketmq
-
-type PullStatus int
-
-const (
-	Found PullStatus = iota
-	NoNewMsg
-	NoMatchedMsg
-	OffsetIllegal
-)
