--- rocketmq-go/remoting_header.go
@@ -1,6 +1,0 @@
-package rocketmq
-
-// TODO to complete
-type CommandCustomHeader interface {
-	checkFields() error
-}
